Changed position type to size_t

+++ include/Ant.h
@@ -6,7 +6,7 @@ class Ant
 private:
     std::vector<size_t> city_order;
     double path_length = 0;
-    uint32_t position = 0;
+    size_t position = 0;
     const size_t ant_index;
     std::vector<bool> is_visited;
 
@@ -22,11 +22,11 @@ public:
     void extendPath(double length) {path_length += length;};
     double getPathLength() const {return path_length;};
 
-    uint32_t getPosition() const {return position;}
-    void setPosition(uint32_t position);
+    size_t getPosition() const {return position;}
+    void setPosition(size_t position);
 
     size_t getIndex() const {return ant_index;}
-    bool isVisited(uint32_t position) const {return is_visited[position];}
+    bool isVisited(size_t position) const {return is_visited[position];}
 
     bool hasEnded();
